Updated ElasticSearch configuration
Updated the elastic search configuration to reflect the new metadata
information, also moved ElasticSearchConfig.cs to AppStart to initialise
the ElasticSearch index on Application start.
