+++ AvaloqDocu/Global.asax.cs
@@ -18,6 +18,7 @@ namespace AvaloqDocu
             GlobalConfiguration.Configure(WebApiConfig.Register);
             RouteConfig.RegisterRoutes(RouteTable.Routes);
             BundleConfig.RegisterBundles(BundleTable.Bundles);
+            ElasticSearchConfig.CreateDefaultIndex();
         }
     }
 }
